move train stats out of net
